windows: Improve hotplug reliability
The device may not be ready at the time of DBT_DEVICEARRIVAL, i.e.
drivers for child device nodes may not have been applied yet.

To deal with this we wait for DBT_DEVNODES_CHANGED. This is however
ambiguous in case of a DBT_DEVICEREMOVECOMPLETE also happening, so we
defer refreshing for a bit longer when such an event is pending.

Co-authored-by: Håvard Sørbø <havard@hsorbo.no>